requirements.txt
# SharePoint integration
Office365-REST-Python-Client==2.5.3

# Excel processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# For older Excel formats if needed
odfpy==1.4.1
